==> 2_Data cleansing.ipynb <==
import numpy as np
import pandas as pd
import math
import sklearn
import sklearn.preprocessing
import datetime
pd.options.mode.chained_assignment = None  

#import dataframe
df = pd.read_csv("/data/workspace_files/merged.csv")
df["eom"] = pd.to_datetime(df["eom"])    #convert to date format
df.head()

df_copy = df.copy()

### Treating missing values

Only the independent variables with less than 25 per cent missing values are retained. The remaining missing values are replaced with the with the cross-sectional median at each month for each stock in line with Gu, Kelly, and Xiu (2020).

- Drop observations where the variable of interest is missing

df = df.dropna( how='any',subset=["ret_excess"])

- Inspect variables with the most missing values:

pd.set_option('display.max_rows',None)
df.isna().sum().sort_values(ascending = False)

- Drop the variables with more than 25% missing values

limitPer = int(len(df) * .75)
df = df.dropna(thresh=limitPer, axis=1)

- Inspect remaining variables:

pd.set_option('display.max_rows',None)
df.isna().sum().sort_values(ascending = False)

- Replace the missing valuues with the cross-sectional median at each month

df = df.fillna(df.groupby('eom').transform('median'))

-  Drop all observations for which there are still missing values

df.dropna(inplace=True)

- There are 2,845,386 observations left

print(len(df))

pd.set_option('display.max_rows',None)
df.isna().sum().sort_values(ascending = False)

## Explore the data before treating 

df = df.sort_values(by=['eom', "permno"])

#the new observation period contains 18 years (2016 Months) form 01.2003 to 12.2020 
print(df['eom'].nunique())

#There are 1855 unique german stocks
print(df['permno'].nunique())

#Inspect variable types
df.info(verbose=True)

### Treating outliers

All monthly firm characteristics are winsorized at the 1% and 99% levels to ensure that the results are insensitive to outliers. In contrast to Gu, Kelly, and Xiu (2020), the dependent variable is also winsorized

df.dtypes

df['date'] = pd.to_datetime(df["date"])

variables = df.columns[~df.columns.isin(["eom",'date','permno'])].tolist()
df[variables] = df[variables].apply(lambda x: x.clip(*x.quantile([0.01, 0.99])))

#Save dataset:
df.to_csv('cleaned.csv', index = False)
