==> 4.1_OLS.ipynb <==
#Import packages
import numpy as np
import pandas as pd
import math
import sklearn
import sklearn.preprocessing
import datetime
import os
import matplotlib.pyplot as plt
from sklearn.linear_model import HuberRegressor
import warnings
warnings.simplefilter(action='ignore', category=Warning)
import seaborn as sns
sns.set()
pd.options.mode.chained_assignment = None  # default='warn'


#Program that executes the rolling window train-val-test split 
%run /data/workspace_files/TimeBasedCV.ipynb

#Read in the cleansed and winsorised data
df = pd.read_csv('/data/workspace_files/cleaned.csv', parse_dates =["eom"])
df.drop('date', axis=1, inplace=True)
df['eom'] = pd.to_datetime(df['eom'])
#Convert to float 32 (format needed for the most ML models)
cols_to_front = ['permno', 'eom', 'ret_excess']
df = df[cols_to_front + [c for c in df.columns if c not in cols_to_front]]
df[df.columns[3:]] = df[df.columns[3:]].astype('float32')
#Sort observations by date and stock id
df = df.sort_values(by = ['eom', 'permno'], ascending = True)
df.head()

### Rank standardization  in inreval [-1,1]

A rank standardization procedure following Kelly, Pruitt, and Su (2019) and Freyberger, Neuhierl, and Weber (2020) is employed. The stock characteristics are ranked month-by-month cross-sectionally and the ranks are mapped into the [-1,1] interval, thereby transforming features into a uniform distribution and increasing the insensitivity to outliers

#Standardize all independent variables
features = df.columns[~df.columns.isin(["eom",'permno',"ret_excess"])].tolist()

#Ranks in [0,1]interval
df[features]=df.groupby("eom")[features].rank(pct=True)

#Multiply by 2 and substract 1 to get ranks in interval [-1,1] 
df[features] = 2*df[features] - 1

df["permno2"] = df["permno"].copy()
df["eom2"] = df["eom"].copy()
df = df.set_index(['eom2','permno2'])
df.head()

# Linear Regression with Huber loss function

## All firms

# Define the splitting scheme:
# Rolling window  train-val-test split
# The rolling window is set to 96 months (8 years)
# 5 years for training, 2 years for validation, and 1 year for out-of-samlpe testing
tscv = TimeBasedCV(train_period=60,
                   val_period=24,
                   test_period=12,
                   freq='months')

#define the idependent variables (X) and the dependent variable (y)
features = df.columns[~df.columns.isin(['eom', 'permno',"ret_excess"])].tolist()
X = df[features]
y = df["ret_excess"]

#Empty containers to save results from each window
predictions = []
y_test_list =[]
dates = []
dic_r2_all = {}


for train_index, val_index, test_index in tscv.split(df, first_split_date= datetime.date(1965,12,31), second_split_date= datetime.date(2020,1,31), date_column="eom"):

    X_train   = X.loc[train_index]
    y_train = y.loc[train_index]
    
    X_val   = X.loc[val_index]
    y_val = y.loc[val_index]

    X_test    = X.loc[test_index]
    y_test  = y.loc[test_index]
    
    #OLS regression with huber loss function
    reg_huber = HuberRegressor(max_iter=1000)
    #No hyperparameters in OLS --> use train and validation set for training
    reg_huber.fit(np.concatenate((X_train, X_val)), np.concatenate((y_train, y_val)))
    #Use test set to generate predictions
    preds = reg_huber.predict(X_test) 
    #Save predictions, dates and the true values of the dependent variable to list  
    predictions.append(preds)
    dates.append(y_test.index)
    y_test_list.append(y_test)
    
    #Calculate OOS model performance the for current window
    r2 = 1-sum(pow(y_test-preds,2))/sum(pow(y_test,2))
    #Save OOS model performance and the respective month to dictionary
    dic_r2_all["r2." + str(y_test.index)] = r2
   
    
    
#Concatenate to get results over the whole OOS test period (Jan 2010-Dec 2019)
predictions_all= np.concatenate(predictions, axis=0)
y_test_list_all= np.concatenate(y_test_list, axis=0) 
dates_all= np.concatenate(dates, axis=0)

#Calculate OOS model performance over the entire test period in line with Gu et al (2020)
R2OOS_LR = 1-sum(pow(y_test_list_all-predictions_all,2))/sum(pow(y_test_list_all,2))
print("R2OOS Hubber Regression: ", R2OOS_LR)

#Generate a results dataframe containing the model predictions (yhat) and the true values (y_true) 
#of the dependent variable for each stock in each month

yhat = predictions_all.tolist()
y_true = y_test_list_all.tolist()
i = dates_all.tolist()

results = pd.DataFrame(
    {'identifier': i,
     'yhat': yhat,
     'y_true': y_true
    })


results['date'] = results['identifier'].str[0]
results['id'] = results['identifier'].str[1]
results.drop(["identifier"],axis = 1, inplace=True)
results['date'] = pd.to_datetime(results['date'], format='%Y-%m-%d')
results['MonthYear'] = results['date'].dt.to_period('M')
results = results.sort_values(by = ['date', 'id'], ascending = True)
results = results.set_index(['MonthYear','id'])
results.head()

# Save results to use for Diebold-Mariano test
ols= results.reset_index()
ols.to_csv('ols.csv', index = False)
